adding counting price for stocks

+++ FXTracker/src/main/java/com/FXTracker/DTO/WalletDto.java
@@ -19,5 +19,5 @@ public class WalletDto {
     @NotNull
     private String userId;
     private String currency;
-    private Float balance;
+    private double balance;
 }
